updated the markdown generating function to return the full layout of the readme. Removed one question from the questions array for being redundant. moved the questions array into the utils folder and now importing it into index.js
